backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
langchain>=0.1.0
langgraph>=0.0.15
openai>=1.12.0
pydantic>=2.0.0
aiohttp>=3.8.0
pytest>=7.0.0
pytest-asyncio>=0.23.0
langchain-openai>=0.0.5 